+++ FredBoat/src/test/java/fredboat/command/admin/TestCommandTest.java
@@ -1,13 +1,9 @@
 package fredboat.command.admin;
 
 import fredboat.Config;
+import fredboat.FakeContext;
 import fredboat.ProvideJDASingleton;
 import fredboat.db.DatabaseManager;
-import fredboat.messaging.internal.Context;
-import net.dv8tion.jda.core.entities.Guild;
-import net.dv8tion.jda.core.entities.Member;
-import net.dv8tion.jda.core.entities.TextChannel;
-import net.dv8tion.jda.core.entities.User;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Assumptions;
@@ -57,38 +53,5 @@ class TestCommandTest extends ProvideJDASingleton {
             dbm.shutdown();
         }
         bumpPassedTests();
-    }
-
-    static class FakeContext extends Context {
-
-        private final TextChannel channel;
-        private final Member member;
-        private final Guild guild;
-
-        public FakeContext(TextChannel channel, Member member, Guild guild) {
-            this.channel = channel;
-            this.member = member;
-            this.guild = guild;
-        }
-
-        @Override
-        public TextChannel getTextChannel() {
-            return channel;
-        }
-
-        @Override
-        public Guild getGuild() {
-            return guild;
-        }
-
-        @Override
-        public Member getMember() {
-            return member;
-        }
-
-        @Override
-        public User getUser() {
-            return getMember().getUser();
-        }
     }
 }
